# docs/04-api/openapi.yaml
openapi: 3.1.0
info:
  title: Financial Research Agent API
  version: 1.0.0
  description: Core public contract. Detailed domain behavior lives in docs/.
servers:
  - url: /v1
security:
  - bearerAuth: []
paths:
  /reports:
    get:
      summary: List reports
      parameters:
        - in: query
          name: limit
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
        - in: query
          name: cursor
          schema: {type: string}
        - in: query
          name: q
          schema: {type: string}
        - in: query
          name: status
          schema: {type: string, enum: [DRAFT, RESEARCHING, READY, VERIFIED, ARCHIVED]}
        - in: query
          name: company_id
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Report page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportListResponse'
    post:
      summary: Create report workspace
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateReportRequest'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ReportSummary'}
...
  /reports/{reportId}/sources:
    parameters:
      - $ref: '#/components/parameters/ReportId'
    get:
      summary: List retention-safe source lineage for a report
      parameters:
        - in: query
          name: limit
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
        - in: query
          name: cursor
          schema: {type: string}
      responses:
        '200':
          description: Owner-authorized source page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ReportSourcePage'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ReportId:
      in: path
      name: reportId
      required: true
      schema: {type: string, format: uuid}
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema: {type: string, minLength: 8, maxLength: 128}
  schemas:
...
    CreateReportRequest:
      type: object
      required: [title, subject]
      properties:
        title: {type: string, minLength: 1, maxLength: 180}
        subject:
          type: object
          required: [query]
          properties:
            query: {type: string}
            country_code: {type: string, minLength: 2, maxLength: 2}
            ticker: {type: string}
            domain: {type: string}
        focus: {type: array, items: {type: string}}
        depth: {type: string, enum: [FAST, STANDARD, DEEP], default: STANDARD}
...
    ReportSummary:
      type: object
      required: [report_id, title, status]
      properties:
        report_id: {type: string, format: uuid}
        title: {type: string}
        status: {type: string}
        current_version: {type: integer, nullable: true}
        updated_at: {type: string, format: date-time}
...
    ReportListResponse:
      type: object
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/ReportSummary'}
        next_cursor: {type: string, nullable: true}
...
    SourceSnapshotSummary:
      type: object
      required: [snapshot_id, retrieved_at, content_hash, retention_mode]
      properties:
        snapshot_id: {type: string, format: uuid}
        retrieved_at: {type: string, format: date-time}
        published_at: {type: string, format: date-time, nullable: true}
        content_hash: {type: string}
        retention_mode: {type: string, enum: [FULL_TEXT, EXCERPT_ONLY, METADATA_ONLY, STORAGE_REFERENCE]}
        provider_request_id: {type: string, format: uuid, nullable: true}
    ReportSourceSummary:
      type: object
      required: [source_id, identity_key, publisher, source_type, authority_tier, ownership_relation, latest_snapshot]
      properties:
        source_id: {type: string, format: uuid}
        identity_key: {type: string}
        canonical_url: {type: string, nullable: true}
        publisher: {type: string}
        domain: {type: string, nullable: true}
        source_type: {type: string}
        authority_tier: {type: string}
        ownership_relation: {type: string}
        source_family_id: {type: string, format: uuid, nullable: true}
        source_family_reason: {type: string, nullable: true}
        latest_snapshot: {$ref: '#/components/schemas/SourceSnapshotSummary'}
    ReportSourcePage:
      type: object
      required: [items]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/ReportSourceSummary'}}
        next_cursor: {type: string, nullable: true}